Step17_JDBC Friend 예제파일 추가

+++ Step17_JDBC/src/test/frame/MemberFrame.java
@@ -153,12 +153,16 @@ public class MemberFrame extends JFrame
 				JOptionPane.showMessageDialog(this, "삭제할 row를 선택하세요!");
 				return;
 			}
-			//2. 해당 인덱스의 table row에서 삭제할 회원의 번호를 읽어와서 (model 활용)
-			int num=(int)model.getValueAt(rowIndex, 0);			
-			//3. MemberDao 객체를 이용해서 DB에서 삭제하고			
-			new MemberDao().delete(num);						
-			//4. 새로고침
-			displayMember();
+			
+			int result=JOptionPane.showConfirmDialog(this, "선택된 row를 삭제하시겠습니까?");
+			if(result == JOptionPane.YES_OPTION) {
+				//2. DefaultTableModel에서 해당 인덱스의 table row에서 삭제할 회원의 번호를 읽어와서 (model 활용)
+				int num=(int)model.getValueAt(rowIndex, 0);			
+				//3. MemberDao 객체를 이용해서 DB에서 삭제하고			
+				new MemberDao().delete(num);						
+				//4. 새로고침
+				displayMember();
+			}			
 		}
 	}
 
@@ -169,22 +173,22 @@ public class MemberFrame extends JFrame
 		System.out.println("property name:"+evt.getPropertyName());
 		
 		System.out.println("isEditing:"+table.isEditing());
-		if(table.isEditing()==false) {
-			int rowIndex=table.getSelectedRow();
-				if(rowIndex!=-1) {
-					int num=(int)model.getValueAt(rowIndex, 0);
-					
-					String name=(String)model.getValueAt(rowIndex, 1);
-					String addr=(String)model.getValueAt(rowIndex, 2);
-					MemberDto dto=new MemberDto();
-					dto.setName(name);
-					dto.setAddr(addr);
-					dto.setNum(num);
-					
-					new MemberDao().update(dto);
-					return;					
-				}
-			
-		}
+		
+		//만일 테이블의 수정사항을 DB에 수정반영할 적당한 시점이 되면
+		if(evt.getPropertyName().equals("tableCellEditor") && !table.isEditing()){
+			//현재 선택된 row의 정보를 DB에 수정반영한다.
+			//변화된 값을 읽어와서 DB에 반영한다.
+			//수정된 칼럼에 있는 row 전체의 값을 읽어온다.
+			int selectedIndex=table.getSelectedRow();
+			int num=(int)model.getValueAt(selectedIndex, 0);
+			String name=(String)model.getValueAt(selectedIndex, 1);
+			String addr=(String)model.getValueAt(selectedIndex, 2);
+			//수정할 회원의 정보를 MemberDto 객체에 담고
+			MemberDto dto=new MemberDto(num, name, addr);			
+			//DB에 저장하기
+			new MemberDao().update(dto);
+			//선택된 row clear
+			table.clearSelection();
+		}		
 	}	
 }
